requirements.txt
h5py~=3.1.0
notebook~=6.4.10
numpy~=1.21.4
Pillow~=8.2.0
tensorflow==2.8.0
tensorflow-federated==0.20.0
tensorflow-probability==0.16.0
tensorflow_addons~=0.17.1
pandas~=1.4.2
